fix: Add unauthorized sender revert into done tasks (#50)
As bridge operators list is changed every period, bridge can be run by
candidates when they are not bridge operator. In this case, their transaction to
bridge contract gets reverted with unauthorized sender message. In order not to
spam the failed tasks list, we consider moving these tasks to done tasks with
last error filled with revert message for tracking .